keep javac on JDK 1.5 quiet
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1189657 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/java/org/apache/poi/xslf/model/geom/LineToCommand.java
@@ -39,6 +39,6 @@ public class LineToCommand implements PathCommand {
     public void execute(GeneralPath path, Context ctx){
         double x = ctx.getValue(arg1);
         double y = ctx.getValue(arg2);
-        path.lineTo(x, y);
+        path.lineTo((float)x, (float)y);
     }
 }
